# glotcast-api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/lists/{slug}": {
      "get": {
        "operationId": "ListsController_page_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/episodes/{id}": {
      "get": {
        "operationId": "EpisodesController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID or legacy Strapi documentId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/admin/episodes": {
      "get": {
        "operationId": "AdminEpisodesController_list_v1",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "title contains (accent-insensitive)",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "podcastId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "draft includes scheduled episodes",
            "schema": {
              "type": "string",
              "enum": [
                "published",
                "draft"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodePageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminEpisodesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEpisodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/episodes/{id}": {
      "get": {
        "operationId": "AdminEpisodesController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminEpisodesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEpisodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminEpisodesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/episodes/{id}/publish": {
      "post": {
        "operationId": "AdminEpisodesController_publish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/episodes/{id}/unpublish": {
      "post": {
        "operationId": "AdminEpisodesController_unpublish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/episodes/{id}/levels/{level}": {
      "put": {
        "operationId": "AdminEpisodesController_putLevel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/Level"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutLevelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminEpisodesController_deleteLevel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/Level"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/lists": {
      "get": {
        "operationId": "AdminListsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListsDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminListsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/lists/{id}": {
      "get": {
        "operationId": "AdminListsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminListsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminListsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/lists/{id}/episodes": {
      "put": {
        "operationId": "AdminListsController_setEpisodes_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListEpisodesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "GlotCast API",
    "description": "Podcasts in three English levels: catalog, listening, vocabulary, translation, and the admin content pipeline. The contract is docs/contract-v1.md; `admin`-tagged operations are for the admin panel only.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ListPageDto_Output": {
        "type": "object",
        "properties": {
          "list": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 0
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "id",
              "slug",
              "name",
              "description"
            ],
            "additionalProperties": false
          },
          "episodes": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EpisodeSummary_Output"
                }
              },
              "page": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "pageSize": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": [
              "items",
              "page",
              "pageSize",
              "total",
              "hasMore"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "list",
          "episodes"
        ],
        "additionalProperties": false
      },
...
      "Level": {
        "type": "string",
        "enum": [
          "bg",
          "in",
          "ad"
        ],
        "description": "bg = beginner, in = intermediate, ad = advanced"
      },
...
      "EpisodeDetailDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "podcast": {
            "$ref": "#/components/schemas/PodcastSummary_Output"
          },
          "number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "bannerUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "isPro": {
            "type": "boolean"
          },
          "publishedAt": {
            "type": "string"
          },
          "levels": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "level": {
                  "$ref": "#/components/schemas/Level_Output"
                },
                "durationSec": {
                  "type": "number"
                },
                "description": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "audioUrl": {
                  "type": "string"
                }
              },
              "required": [
                "level",
                "durationSec",
                "description",
                "audioUrl"
              ],
              "additionalProperties": false
            }
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "isFavorite": {
            "type": "boolean"
          },
          "progress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LevelProgress_Output"
            },
            "description": "the caller's per-level progress, [] for none"
          }
        },
        "required": [
          "id",
          "podcast",
          "number",
          "title",
          "coverUrl",
          "bannerUrl",
          "isPro",
          "publishedAt",
          "levels",
          "description",
          "isFavorite",
          "progress"
        ],
        "additionalProperties": false
      },
...
      "AdminEpisodePageDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminEpisode_Output"
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "CreateEpisodeDto": {
        "type": "object",
        "properties": {
          "podcastId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 100000
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ]
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "bannerUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "isPro": {
            "description": "default true",
            "type": "boolean"
          },
          "publishedAt": {
            "description": "schedule (future) or backdate; null = draft. POST …/publish publishes now",
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "podcastId",
          "title"
        ]
      },
...
      "UpdateEpisodeDto": {
        "type": "object",
        "properties": {
          "podcastId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 100000
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ]
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "bannerUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "isPro": {
            "description": "default true",
            "type": "boolean"
          },
          "publishedAt": {
            "description": "schedule (future) or backdate; null = draft. POST …/publish publishes now",
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "PutLevelDto": {
        "type": "object",
        "properties": {
          "audioUrl": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "durationSec": {
            "type": "number",
            "minimum": 0,
            "maximum": 86400
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "transcript": {
            "type": "object",
            "properties": {
              "chunks": {
                "maxItems": 50000,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TranscriptChunk"
                }
              }
            },
            "required": [
              "chunks"
            ]
          }
        },
        "required": [
          "audioUrl",
          "durationSec",
          "transcript"
        ]
      },
...
      "AdminListsDto_Output": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AdminList_Output"
        }
      },
...
      "ListEpisodesDto": {
        "type": "object",
        "properties": {
          "episodeIds": {
            "maxItems": 500,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "description": "the whole list, in order"
          }
        },
        "required": [
          "episodeIds"
        ]
      },
...
    }
  }
}